+++ src/tightdb/table_accessors.hpp
@@ -797,6 +797,7 @@ public:
 
     bool has_search_index() const { return m_table->get_impl()->has_search_index(col_idx); }
     void add_search_index() const { m_table->get_impl()->add_search_index(col_idx); }
+    void remove_search_index() const { m_table->get_impl()->remove_search_index(col_idx); }
 
     BasicTableView<RealTable> get_sorted_view(bool ascending=true) const
     {