+++ release_notes.md
@@ -12,6 +12,7 @@
 
 * Improved performance of advance_read() over commits with string or binary data insertions.
 * Improved performance sorting TableView and LinkView.
+* Added Table::remove_search_index().
 
 -----------